fix: change bigint to float

+++ src/cart-item/entities/cart-item.entity.ts
@@ -43,7 +43,7 @@ export class CartItem {
 
   @Column({
     name: 'subtotal',
-    type: 'bigint',
+    type: 'float',
     nullable: false,
   })
   @ApiProperty({
